Add fixme notes regarding unfixed bugs
These notes serve as reminders about unfixed issues
with the current codebase.

Wine EXE termination:

The current code will need fixing someday. It currently
does a nonsensical scan for all running processes with
".exe" somewhere on the "ps" line. This means that it's
full of false positives, and will find random EXEs that
are running under other Wine prefixes and which have
nothing to do with FFXIV.

RPATH generation:

It needs to be able to handle escaped colons. The current
code can lead to corrupt path extraction.