Stage 2: Fix broken EXE terminator and use lib
The old code was broken. Its intent was to close the Wine
".exe" processes. But it only closed one PID, even though
there are lots processes (such as FFXIV, ACT, XIVlauncher
or the normal launcher, and all internal Wine processes
such as "explorer.exe", etc).

We now use the new PID scanner library, and use a loop to
terminate every ".exe" process. Thanks to the scanner,
we're closing them in reverse order of how they were
started, which will help with reliability since programs
won't be shut down from underneath a higher layer of the
Wine process list.

There still remains the issue that we're closing every
Wine process even when they don't belong to the FFXIX Wine
container that we're trying to patch, but fixing that
requires more thinking and a future solution.