Properly escape all variables
This is the result of a very deep review of all code,
looking for every instance where variables weren't
properly escaped, which would cause issues with
things such as spaces in paths.

The next commit will be deeply linked to this commit,
and they are both required for the next commit to work.