Stage 1: Unescape variables extracted from raw env
This commit requires the previous commit.

A new method of unescaping the raw variables is now used,
which ensures that we can treat the extracted data like any
other variable, without needing to remember that some data
is escaped and some isn't. Everything is now much more
intuitive for code maintenance and future bug avoidance.